Add dashboard layout and separate home layout
