+++ src/app/(dashboardLayout)/dashboard/page.tsx
@@ -1,0 +1,11 @@
+import Link from "next/link";
+
+const page = () => {
+    return (
+        <div>
+            <Link href="/">Home</Link>
+        </div>
+    );
+};
+
+export default page;
